requirements.txt
Django==2.1
django-import-export==1.0.1
openpyxl==2.5.5
psycopg2==2.7.5
pytz==2018.5
boto3==1.8.3